+++ components/charge-page/AddComment.tsx
@@ -5,11 +5,11 @@ import {
     SelectItem,
     SelectTrigger,
     SelectValue,
-} from "./ui/select";
-import { Input } from "./ui/input";
-import { Label } from "./ui/label";
-import { Button } from "./ui/button";
-import { Textarea } from "./ui/textarea";
+} from "components/ui/select";
+import { Input } from "components/ui/input";
+import { Label } from "components/ui/label";
+import { Button } from "components/ui/button";
+import { Textarea } from "components/ui/textarea";
 
 export default function AddComment() {
     return (
