+++ components/charge-page/ChargeCompanyInfo.tsx
@@ -1,5 +1,5 @@
 import { cn } from "@/lib/utils";
-import { Card, CardContent } from "./ui/card";
+import { Card, CardContent } from "components/ui/card";
 
 type CardProps = React.ComponentProps<typeof Card>;
