+++ components/charge-page/CommentCard.tsx
@@ -1,5 +1,5 @@
 import { ThumbsUp, ThumbsDown } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "components/ui/button";
 import {
     Card,
     CardContent,
@@ -7,7 +7,7 @@ import {
     CardFooter,
     CardHeader,
     CardTitle,
-} from "./ui/card";
+} from "components/ui/card";
 
 export default function CommentCard() {
     return (
