+++ components/home/ChargeItem.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 
 type ChargeItemProps = {
     chargeName: string;
