+++ components/home/MostPopularCharges.tsx
@@ -1,14 +1,6 @@
-import {
-    Card,
-    CardContent,
-    CardDescription,
-    CardFooter,
-    CardHeader,
-    CardTitle,
-} from "./ui/card";
-
-import ChargeItem from "./ChargeItem";
 import { db } from "@/lib/db";
+import ChargeItem from "./ChargeItem";
+import { Card, CardHeader, CardTitle, CardContent } from "components/ui/card";
 
 async function getCharges() {
     const res = await db.charge.findMany({
